uapi: convert stat.h
